+++ src/main/java/github/com/jbabe/web/dto/authAccount/AccessAndRefreshToken.java
@@ -10,17 +10,6 @@ import org.springframework.http.ResponseCookie;
 @AllArgsConstructor
 public class AccessAndRefreshToken {
     private String accessToken;
-    private ResponseCookie responseCookie;
     private String refreshToken;
-
-    public AccessAndRefreshToken(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-
-    public AccessAndRefreshToken(String accessToken, ResponseCookie responseCookie) {
-        this.accessToken = accessToken;
-        this.responseCookie = responseCookie;
-    }
 }
