+++ src/main/java/github/com/jbabe/web/dto/authAccount/LoginResponse.java
@@ -1,5 +1,6 @@
 package github.com.jbabe.web.dto.authAccount;
 
+import github.com.jbabe.repository.user.User;
 import lombok.*;
 
 @Getter
@@ -7,10 +8,10 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class SocialLoginResponse {
-    private Integer status;
-    private String message;
+public class LoginResponse {
     private String accessToken;
     private String refreshToken;
     private String email;
+    private String name;
+    private User.Role role;
 }
